refactor(via): rename params mod to path_params
