+++ src/request/params/mod.rs
@@ -1,6 +1,6 @@
 mod decode;
 mod param;
-mod params;
+mod path_params;
 mod query_param;
 mod query_parser;
 
@@ -8,4 +8,4 @@ pub use param::Param;
 pub use query_param::{QueryParam, QueryParamIter};
 
 pub(crate) use decode::{DecodeParam, NoopDecoder, PercentDecoder};
-pub(crate) use params::PathParams;
+pub(crate) use path_params::PathParams;
